add default devices selector
