+++ core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/components/dot-uve-toolbar/components/dot-uve-device-selector/dot-uve-device-selector.component.html
@@ -1,24 +1,10 @@
-<p-button
-    icon="pi pi-desktop"
-    pTooltip="Desktop Preview"
-    styleClass="p-button-text p-button-sm"
-    data-testId="desktop-preview" />
-
-<p-button
-    icon="pi pi-tablet"
-    pTooltip="Tablet Preview"
-    styleClass="p-button-text p-button-sm"
-    data-testId="tablet-preview" />
-
-<p-button
-    icon="pi pi-mobile"
-    pTooltip="Mobile Preview"
-    styleClass="p-button-text p-button-sm"
-    data-testId="mobile-preview" />
-
-<!-- <p-splitButton
-    styleClass="p-button-sm"
-    label="More"
-    [model]="[]"
-    text
-    data-testId="more-devices-preview" /> -->
+@for (device of defaultDevices; track $index) {
+    <p-button
+        [icon]="device.icon"
+        [pTooltip]="device.name | dm"
+        tooltipPosition="bottom"
+        styleClass="p-button-text p-button-sm"
+        [ngClass]="{ 'p-button-success': device.inode === $currentDevice()?.inode }"
+        (click)="onDeviceSelect(device)"
+        [attr.data-testId]="device.inode" />
+}
